requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
twilio==8.10.0
sendgrid==6.11.0
geojson==3.1.0
pydantic==2.5.0
SQLAlchemy==2.0.23